requirements.txt
# Core dependencies
Pillow>=10.0.0  # Python Imaging Library for image processing
numpy>=1.24.0   # Required for array operations and calculations

# Optional but recommended packages
tqdm>=4.65.0    # Progress bars for long-running operations
python-json-logger>=2.0.7  # Better JSON logging format
typing-extensions>=4.7.1   # Enhanced type hinting support

# Development dependencies
pytest>=7.4.0    # Testing framework
pytest-cov>=4.1.0  # Test coverage reporting
black>=23.7.0     # Code formatting
isort>=5.12.0     # Import sorting
flake8>=6.1.0     # Code linting
mypy>=1.5.1       # Static type checking

# Documentation
Sphinx>=7.1.2     # Documentation generator
sphinx-rtd-theme>=1.3.0  # Read the Docs theme
sphinx-autodoc-typehints>=1.24.0  # Type hints in documentation

# Optional format support
imagecodecs>=2023.7.10  # Additional image format support (JPEG2000, WEBP)

# Performance optimization
psutil>=5.9.5     # System resource monitoring

# Web Hosting
streamlit 
pandas
